src/requirements.txt
click==8.1.3
loguru==0.7.0
filelock==3.13.1
gradio==4.24.0
multiprocess==0.70.16
expiringdict==1.2.2
gevent==24.2.1
zerorpc==0.6.3
datadreamer.dev==0.35.0
scikit-learn==1.4.1.post1

# WebArena requirements
-e webarena_benchmark
-r ../webarena_benchmark/requirements.txt